Refactored calculator to return double.

+++ src/main/java/javaclasses/clock/ArrowAngleCalculator.java
@@ -2,6 +2,6 @@ package javaclasses.clock;
 
 public interface ArrowAngleCalculator {
 
-    int calculate(int hours, int minutes);
+    double calculate(int hours, int minutes);
 
 }
